requirements.txt
aniso8601==8.0.0
Click==7.0
dlib==19.18.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
opencv-contrib-python==4.1.1.26
openface==0.2.1
Pillow==6.2.0
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.9
Werkzeug==0.16.0
